requirements.txt
aiohttp>=3.9.0
openai>=1.0.0
opensearch-py>=2.0.0
pymupdf>=1.24.0
python-dotenv>=1.0.0
motor>=3.3.0
pymongo>=4.6.0
python-magic>=0.4.27
